Add files via upload
Using Python, I'm trying to create and run a program that will calculate the purchase of three different fruits: apple, banana, and orange. Based on the quantity of each fruit given by the user, it should calculate the subtotal and change due, if need be. These are random prices that I thought on the top of my head, and do not reflect the current market for such fruit. With that being said: Owner's cash at hand = $20.00 Apple = $0.75 Banana = $0.80 Orange = $0.99

Enjoy!

And please let me know if you have any suggestions! I'm still learning Python, so your input (Python pun) would be really helpful!